refactor!: webpack integration
move massage templates to separate file

move sendMessage to separate file

BREAKING CHANGE
