+++ package.json
@@ -4,6 +4,7 @@
   "description": "",
   "private": true,
   "dependencies": {
+    "@babel/polyfill": "^7.8.3",
     "cookie-session": "^1.4.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
@@ -20,13 +21,14 @@
     "env-cmd": "^10.0.1",
     "nodemon": "^2.0.2",
     "webpack": "^4.41.5",
-    "webpack-cli": "^3.3.10"
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.10.1"
   },
   "scripts": {
     "dev": "env-cmd -f ./config/config.env nodemon src/server.js -e js,css,hbs,html",
-    "start": "env-cmd -f ./config/config.env node src/server.js",
-    "build": "webpack",
-    "dev-pack": "npx webpack --config webpack.config.js"
+    "start": "webpack --mode development --watch",
+    "build": "webpack --mode production",
+    "server": "webpack-dev-server --inline --hot"
   },
   "browserslist": [
     "last 1 version",
