+++ public/js/chat.js
@@ -1,6 +1,8 @@
-import _ from 'lodash'
+const _ = require('lodash')
+require('@babel/polyfill') // to use async functions with webpack
 
-console.log('Success')
+const { userTemplate, assistantTemplate } = require('./messageTemplates')
+const { sendMessageToAssistant } = require('./sendMessage')
 
 const $messageForm = document.querySelector('#message-form')
 const $messageFormInput = $messageForm.querySelector('input')
@@ -15,12 +17,12 @@ $messageForm.addEventListener('submit', async e => {
     $messageFormButton.setAttribute('disabled', 'disabled')
     const message = e.target.message.value
 
-    const userOutput = userMessage(message)
+    const userOutput = userTemplate(message)
     updateChatDisplay(userOutput)
 
     const { assistant_response, intent } = await sendMessageToAssistant(message)
-    const assistantOutput = assistMessage(assistant_response)
-    // TODO: Block the option to send empty message
+    const assistantOutput = assistantTemplate(assistant_response)
+    // TODO: Rectrict the option to send empty message
     updateChatDisplay(assistantOutput)
     setButtonAndView(intent)
 })
@@ -42,65 +44,7 @@ const updateChatDisplay = messageTemplate => {
     $chatDisplay.innerHTML += messageTemplate
 }
 
-const sendMessageToAssistant = async userInput => {
-    const formData = { message: userInput }
-
-    const res = await fetch('/api/message', {
-        method: 'post',
-        headers: {
-            'Content-type': 'application/json; charset=UTF-8'
-        },
-        body: JSON.stringify(formData)
-    })
-    let data = await res.json()
-    const assistant_response = data.assistant_response
-    const intent = data.intent
-
-    return { assistant_response, intent }
-}
-
-const userMessage = msg_content => {
-    const time = getCurrentTime()
-    return `
-        <div class="ui grid">
-            <div class="six wide column">
-                <div class="ui blue message user-input msg user">
-                    <p>${msg_content}</p>
-                </div>
-            </div>
-            <div class="ten wide column time-display">
-                    <p class="time">${time}</p>
-            </div>
-        </div>
-`
-}
-const assistMessage = msg_content => {
-    const time = getCurrentTime()
-
-    return `
-    <hr class="break">
-        <div class="ui grid">
-            <div class="ten wide column time-display">
-                <p class="time">${time}</p>
-            </div>
-        <div class="six wide column">
-            <div class="ui success message user-input msg assist">
-                <p>${msg_content}</p>
-            </div>
-        </div>
-    </div>
-    `
-}
-
 function scrollToBottom() {
     $chatDisplay.scrollTop =
         $chatDisplay.scrollHeight - $chatDisplay.clientHeight
-}
-
-const getCurrentTime = () => {
-    const this_time = new Date()
-    const this_hour = this_time.getHours()
-    const this_Minutes = this_time.getMinutes()
-    const time = `${this_hour}:${this_Minutes}`
-    return time
 }
