+++ src/app.js
@@ -17,7 +17,7 @@ hbs.registerPartials(partialPath)
 app.use(cors())
 app.set('trust proxy', 1) // trust first proxy
 
-app.use(express.static(publicDirectoryPath)) // load UI from public folder
+app.use(express.static(publicDirectoryPath))
 app.use(express.json())
 app.use(express.urlencoded({ extended: true }))
