+++ templates/views/chat.hbs
@@ -28,7 +28,7 @@
 
 
     </div>
-    <script src="main.js"></script>
+    <script src="/js/chat.bundle.js"></script>
 </body>
 
 </html>
